add WeekDay.c, DaysBetweenDates.c, date.h
In order to create the calendar with all functionalities intended, it'll be an easier path if separate files are made first and combine them all later on.

date.h: the main library on this project it has lots of useful functions for the other programs

WeekDay.c: It simply shows which week day it was on a certain date

DaysBetweenDates.c: It shows how many days there are between 2 dates.